+++ app/lib/bk/compat/models/steps/input.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module BK
+  module Compat
+    # input steps are almost the same as block steps (difference is dependency semantics)
+    class InputStep < BlockStep
+      def to_h
+        super.tap do |h|
+          h[:input] = h.delete(:block)
+        end
+      end
+    end
+  end
+end
