+++ app/lib/bk/compat/parsers/bitbucket/variables.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require_relative '../../pipeline/step'
 require_relative '../../models/plugin'
+require_relative '../../models/steps/input'
 
 module BK
   module Compat
